Patchset4/4
The AIHard class now correctly extends the newly created
AIDifficulty superclass. The traces of "getType()" and "getCommand()"
methods were also removed from this class since they were pulled up to
AIDifficulty on the previous patchset. As a result, the attributes
"type" and "command" are directly assigned of the appropriate
difficulty for this class.